remove the pkg/config/env.go because it is unused or obsolete. Move the only useful thing from that file - SPIAccessTokenLinkLabel constant to a new file in controllers package.
